Renaming lightning module to have year and source.
